:recycle: refactor: AES256 암호화 클래스를 전역처리하지 않도록 수정
