--- semo-core/src/main/java/sandbox/semo/domain/common/crypto/ApplicationContextUtil.java
@@ -1,21 +1,0 @@
-package sandbox.semo.domain.common.crypto;
-
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.stereotype.Component;
-
-@Component
-public class ApplicationContextUtil implements ApplicationContextAware {
-
-    private static ApplicationContext context;
-
-    @Override
-    public void setApplicationContext(ApplicationContext applicationContext) {
-        context = applicationContext;
-    }
-
-    public static <T> T getBean(Class<T> beanClass) {
-        return context.getBean(beanClass);
-    }
-
-}
